Remove 'volt plugconf' command
