+++ main.go
@@ -32,8 +32,6 @@ func Main() int {
 		return cmd.List(os.Args[2:])
 	case "profile":
 		return cmd.Profile(os.Args[2:])
-	case "plugconf":
-		return cmd.Plugconf(os.Args[2:])
 	case "rebuild":
 		return cmd.Rebuild(os.Args[2:])
 	case "version":
